refactor: unify code to run.py, tidy up overriding functions
