+++ evaluation/eval_openai.py
@@ -1,3 +1,6 @@
+# TODO: Refactor to match other models way of running
+# TODO: add dependencies to req.txt (if any)
+
 from openai import OpenAI
 client = OpenAI()
